Removed useless files
Haven't looked at this project in a while. Went back to check out what
I had originally put in here. Realized that is was full of garbage with
the exception of the rss-fetch (I still use that one). So I cleaned things
up a bit and got rid of the junk.

+ Removed abs_dir.rb
+ Removed git_commit_prepare.rb
+ Removed svn_production_prepare.rb
+ Updated README with a simple description
+ Added hash-bang to top of rss-previewer script